Moving bar_to_abar circuit to its file https://findora.atlassian.net/browse/FO-963
