+++ api/src/anon_xfr/circuits.rs
@@ -1,21 +1,13 @@
 use crate::anon_xfr::structs::{BlindFactor, Commitment, MTNode, MTPath, Nullifier};
-use num_bigint::BigUint;
 use zei_algebra::{
     bls12_381::BLSScalar,
     jubjub::{JubjubPoint, JubjubScalar},
     ops::*,
-    ristretto::RistrettoScalar,
-    traits::{Group, Scalar},
+    traits::Group,
     One, Zero,
 };
 use zei_crypto::basic::rescue::RescueInstance;
-use zei_crypto::{
-    field_simulation::{SimFr, BIT_PER_LIMB, NUM_OF_LIMBS},
-    pc_eq_rescue_split_verifier_zk_part::{NonZKState, ZKPartProof},
-};
-use zei_plonk::plonk::constraint_system::{
-    field_simulation::SimFrVar, rescue::StateVar, TurboConstraintSystem, VarIndex,
-};
+use zei_plonk::plonk::constraint_system::{rescue::StateVar, TurboConstraintSystem, VarIndex};
 
 pub type TurboPlonkCS = TurboConstraintSystem<BLSScalar>;
 
@@ -309,237 +301,6 @@ pub(crate) fn build_multi_xfr_cs(
 
     cs.prepare_io_variable(hash_var);
     cs.prepare_io_variable(non_malleability_tag_var);
-
-    // pad the number of constraints to power of two
-    cs.pad();
-
-    let n_constraints = cs.size;
-    (cs, n_constraints)
-}
-
-/// Returns the constraint system (and associated number of constraints) for equality of values
-/// in a Pedersen commitment and a Rescue commitment.
-pub(crate) fn build_eq_committed_vals_cs(
-    amount: BLSScalar,
-    asset_type: BLSScalar,
-    blind_hash: BLSScalar,
-    pubkey_x: BLSScalar,
-    proof: &ZKPartProof,
-    non_zk_state: &NonZKState,
-    beta: &RistrettoScalar,
-    lambda: &RistrettoScalar,
-) -> (TurboPlonkCS, usize) {
-    let mut cs = TurboConstraintSystem::new();
-    let zero_var = cs.zero_var();
-
-    let zero = BLSScalar::zero();
-    let one = BLSScalar::one();
-    let step_1 = BLSScalar::from(&BigUint::one().shl(BIT_PER_LIMB));
-    let step_2 = BLSScalar::from(&BigUint::one().shl(BIT_PER_LIMB * 2));
-    let step_3 = BLSScalar::from(&BigUint::one().shl(BIT_PER_LIMB * 3));
-    let step_4 = BLSScalar::from(&BigUint::one().shl(BIT_PER_LIMB * 4));
-    let step_5 = BLSScalar::from(&BigUint::one().shl(BIT_PER_LIMB * 5));
-
-    // 1. Input Ristretto commitment data
-    let amount_var = cs.new_variable(amount);
-    let at_var = cs.new_variable(asset_type);
-    let blind_hash_var = cs.new_variable(blind_hash);
-    let pubkey_x_var = cs.new_variable(pubkey_x);
-
-    // 2. Input witness x, y, a, b, r, public input comm, beta, s1, s2
-    let x_sim_fr = SimFr::from(&BigUint::from_bytes_le(&non_zk_state.x.to_bytes()));
-    let y_sim_fr = SimFr::from(&BigUint::from_bytes_le(&non_zk_state.y.to_bytes()));
-    let a_sim_fr = SimFr::from(&BigUint::from_bytes_le(&non_zk_state.a.to_bytes()));
-    let b_sim_fr = SimFr::from(&BigUint::from_bytes_le(&non_zk_state.b.to_bytes()));
-    let comm = proof.non_zk_part_state_commitment;
-    let r = non_zk_state.r;
-
-    let beta_sim_fr = SimFr::from(&BigUint::from_bytes_le(&beta.to_bytes()));
-    let lambda_sim_fr = SimFr::from(&BigUint::from_bytes_le(&lambda.to_bytes()));
-
-    let beta_lambda = *beta * lambda;
-    let beta_lambda_sim_fr = SimFr::from(&BigUint::from_bytes_le(&beta_lambda.to_bytes()));
-
-    let s1_plus_lambda_s2 = proof.s_1 + proof.s_2 * lambda;
-    let s1_plus_lambda_s2_sim_fr =
-        SimFr::from(&BigUint::from_bytes_le(&s1_plus_lambda_s2.to_bytes()));
-
-    let x_sim_fr_var = SimFrVar::alloc_witness_bounded_total_bits(&mut cs, &x_sim_fr, 64);
-    let y_sim_fr_var = SimFrVar::alloc_witness_bounded_total_bits(&mut cs, &y_sim_fr, 240);
-    let a_sim_fr_var = SimFrVar::alloc_witness(&mut cs, &a_sim_fr);
-    let b_sim_fr_var = SimFrVar::alloc_witness(&mut cs, &b_sim_fr);
-    let comm_var = cs.new_variable(comm);
-    let r_var = cs.new_variable(r);
-    let beta_sim_fr_var = SimFrVar::alloc_input(&mut cs, &beta_sim_fr);
-    let lambda_sim_fr_var = SimFrVar::alloc_input(&mut cs, &lambda_sim_fr);
-    let beta_lambda_sim_fr_var = SimFrVar::alloc_input(&mut cs, &beta_lambda_sim_fr);
-    let s1_plus_lambda_s2_sim_fr_var = SimFrVar::alloc_input(&mut cs, &s1_plus_lambda_s2_sim_fr);
-
-    // 3. Merge the limbs for x, y, a, b
-    let mut all_limbs = Vec::with_capacity(4 * NUM_OF_LIMBS);
-    all_limbs.extend_from_slice(&x_sim_fr.limbs);
-    all_limbs.extend_from_slice(&y_sim_fr.limbs);
-    all_limbs.extend_from_slice(&a_sim_fr.limbs);
-    all_limbs.extend_from_slice(&b_sim_fr.limbs);
-
-    let mut all_limbs_var = Vec::with_capacity(4 * NUM_OF_LIMBS);
-    all_limbs_var.extend_from_slice(&x_sim_fr_var.var);
-    all_limbs_var.extend_from_slice(&y_sim_fr_var.var);
-    all_limbs_var.extend_from_slice(&a_sim_fr_var.var);
-    all_limbs_var.extend_from_slice(&b_sim_fr_var.var);
-
-    let mut compressed_limbs = Vec::with_capacity(5);
-    let mut compressed_limbs_var = Vec::with_capacity(5);
-    for (limbs, limbs_var) in all_limbs.chunks(5).zip(all_limbs_var.chunks(5)) {
-        let mut sum = BigUint::zero();
-        for (i, limb) in limbs.iter().enumerate() {
-            sum.add_assign(<&BLSScalar as Into<BigUint>>::into(limb).shl(BIT_PER_LIMB * i));
-        }
-        compressed_limbs.push(BLSScalar::from(&sum));
-
-        let mut sum_var = {
-            let first_var = *limbs_var.get(0).unwrap_or(&zero_var);
-            let second_var = *limbs_var.get(1).unwrap_or(&zero_var);
-            let third_var = *limbs_var.get(2).unwrap_or(&zero_var);
-            let fourth_var = *limbs_var.get(3).unwrap_or(&zero_var);
-
-            cs.linear_combine(
-                &[first_var, second_var, third_var, fourth_var],
-                one,
-                step_1,
-                step_2,
-                step_3,
-            )
-        };
-
-        if limbs.len() == 5 {
-            let fifth_var = *limbs_var.get(4).unwrap_or(&zero_var);
-            sum_var = cs.linear_combine(
-                &[sum_var, fifth_var, zero_var, zero_var],
-                one,
-                step_4,
-                zero,
-                zero,
-            );
-        }
-
-        compressed_limbs_var.push(sum_var);
-    }
-
-    // 4. Open the non-ZK verifier state
-    {
-        let h1_var = cs.rescue_hash(&StateVar::new([
-            compressed_limbs_var[0],
-            compressed_limbs_var[1],
-            compressed_limbs_var[2],
-            compressed_limbs_var[3],
-        ]))[0];
-
-        let h2_var = cs.rescue_hash(&StateVar::new([
-            h1_var,
-            compressed_limbs_var[4],
-            r_var,
-            zero_var,
-        ]))[0];
-        cs.equal(h2_var, comm_var);
-    }
-
-    // 5. Perform the check in field simulation
-    {
-        let beta_x_sim_fr_mul_var = beta_sim_fr_var.mul(&mut cs, &x_sim_fr_var);
-        let beta_lambda_y_sim_fr_mul_var = beta_lambda_sim_fr_var.mul(&mut cs, &y_sim_fr_var);
-        let lambda_b_sim_fr_mul_var = lambda_sim_fr_var.mul(&mut cs, &b_sim_fr_var);
-
-        let mut rhs = beta_x_sim_fr_mul_var.add(&mut cs, &beta_lambda_y_sim_fr_mul_var);
-        rhs = rhs.add(&mut cs, &lambda_b_sim_fr_mul_var);
-
-        let s1_plus_lambda_s2_minus_a_sim_fr_var =
-            s1_plus_lambda_s2_sim_fr_var.sub(&mut cs, &a_sim_fr_var);
-
-        let eqn = rhs.sub(&mut cs, &s1_plus_lambda_s2_minus_a_sim_fr_var);
-        eqn.enforce_zero(&mut cs);
-    }
-
-    // 6. Check x = amount_var and y = at_var
-    {
-        let mut x_in_bls12_381 = cs.linear_combine(
-            &[
-                x_sim_fr_var.var[0],
-                x_sim_fr_var.var[1],
-                x_sim_fr_var.var[2],
-                x_sim_fr_var.var[3],
-            ],
-            one,
-            step_1,
-            step_2,
-            step_3,
-        );
-        x_in_bls12_381 = cs.linear_combine(
-            &[
-                x_in_bls12_381,
-                x_sim_fr_var.var[4],
-                x_sim_fr_var.var[5],
-                zero_var,
-            ],
-            one,
-            step_4,
-            step_5,
-            zero,
-        );
-
-        let mut y_in_bls12_381 = cs.linear_combine(
-            &[
-                y_sim_fr_var.var[0],
-                y_sim_fr_var.var[1],
-                y_sim_fr_var.var[2],
-                y_sim_fr_var.var[3],
-            ],
-            one,
-            step_1,
-            step_2,
-            step_3,
-        );
-        y_in_bls12_381 = cs.linear_combine(
-            &[
-                y_in_bls12_381,
-                y_sim_fr_var.var[4],
-                y_sim_fr_var.var[5],
-                zero_var,
-            ],
-            one,
-            step_4,
-            step_5,
-            zero,
-        );
-
-        cs.equal(x_in_bls12_381, amount_var);
-        cs.equal(y_in_bls12_381, at_var);
-    }
-
-    // 7. Rescue commitment
-    let rescue_comm_var = cs.rescue_hash(&StateVar::new([
-        blind_hash_var,
-        amount_var,
-        at_var,
-        pubkey_x_var,
-    ]))[0];
-
-    // prepare public inputs
-    cs.prepare_io_variable(rescue_comm_var);
-    cs.prepare_io_variable(comm_var);
-
-    for i in 0..NUM_OF_LIMBS {
-        cs.prepare_io_variable(beta_sim_fr_var.var[i]);
-    }
-    for i in 0..NUM_OF_LIMBS {
-        cs.prepare_io_variable(lambda_sim_fr_var.var[i]);
-    }
-    for i in 0..NUM_OF_LIMBS {
-        cs.prepare_io_variable(beta_lambda_sim_fr_var.var[i]);
-    }
-    for i in 0..NUM_OF_LIMBS {
-        cs.prepare_io_variable(s1_plus_lambda_s2_sim_fr_var.var[i]);
-    }
 
     // pad the number of constraints to power of two
     cs.pad();
@@ -892,8 +653,6 @@ pub(crate) mod tests {
     use ruc::*;
     use zei_algebra::{bls12_381::BLSScalar, traits::Scalar};
     use zei_crypto::basic::rescue::RescueInstance;
-    use zei_crypto::basic::ristretto_pedersen_comm::RistrettoPedersenCommitment;
-    use zei_crypto::pc_eq_rescue_split_verifier_zk_part::prove_pc_eq_rescue_external;
     use zei_plonk::plonk::constraint_system::{ecc::Point, TurboConstraintSystem};
 
     pub(crate) fn new_multi_xfr_witness_for_test(
@@ -1541,78 +1300,6 @@ pub(crate) mod tests {
         asset_mixing(&mut cs, &inputs, &outputs, fee_type, &fee_calculating_func);
         let witness = cs.get_and_clear_witness();
         assert!(cs.verify_witness(&witness, &[]).is_err());
-    }
-
-    #[test]
-    fn test_eq_committed_vals_cs() {
-        let mut rng = ChaChaRng::from_seed([0u8; 32]);
-        let pc_gens = RistrettoPedersenCommitment::default();
-
-        // 1. compute the parameters
-        let amount = BLSScalar::from(71u32);
-        let asset_type = BLSScalar::from(52u32);
-
-        let x = RistrettoScalar::from_bytes(&amount.to_bytes()).unwrap();
-        let y = RistrettoScalar::from_bytes(&asset_type.to_bytes()).unwrap();
-
-        let gamma = RistrettoScalar::random(&mut rng);
-        let delta = RistrettoScalar::random(&mut rng);
-
-        let point_p = pc_gens.commit(x, gamma);
-        let point_q = pc_gens.commit(y, delta);
-
-        let z_randomizer = BLSScalar::random(&mut rng);
-        let z_instance = RescueInstance::<BLSScalar>::new();
-
-        let x_in_bls12_381 = BLSScalar::from(&BigUint::from_bytes_le(&x.to_bytes()));
-        let y_in_bls12_381 = BLSScalar::from(&BigUint::from_bytes_le(&y.to_bytes()));
-
-        let pubkey_x = BLSScalar::random(&mut rng);
-
-        let z = z_instance.rescue(&[z_randomizer, x_in_bls12_381, y_in_bls12_381, pubkey_x])[0];
-
-        // 2. compute the ZK part of the proof
-        let (proof, non_zk_state, beta, lambda) = prove_pc_eq_rescue_external(
-            &mut rng, &x, &gamma, &y, &delta, &pc_gens, &point_p, &point_q, &z,
-        )
-        .unwrap();
-
-        // compute cs
-        let (mut cs, _) = build_eq_committed_vals_cs(
-            amount,
-            asset_type,
-            z_randomizer,
-            pubkey_x,
-            &proof,
-            &non_zk_state,
-            &beta,
-            &lambda,
-        );
-        let witness = cs.get_and_clear_witness();
-
-        let mut online_inputs = Vec::with_capacity(2 + 3 * NUM_OF_LIMBS);
-        online_inputs.push(z);
-        online_inputs.push(proof.non_zk_part_state_commitment);
-
-        let beta_sim_fr = SimFr::from(&BigUint::from_bytes_le(&beta.to_bytes()));
-        let lambda_sim_fr = SimFr::from(&BigUint::from_bytes_le(&lambda.to_bytes()));
-
-        let beta_lambda = beta * &lambda;
-        let beta_lambda_sim_fr = SimFr::from(&BigUint::from_bytes_le(&beta_lambda.to_bytes()));
-
-        let s1_plus_lambda_s2 = proof.s_1 + proof.s_2 * lambda;
-        let s1_plus_lambda_s2_sim_fr =
-            SimFr::from(&BigUint::from_bytes_le(&s1_plus_lambda_s2.to_bytes()));
-
-        online_inputs.extend_from_slice(&beta_sim_fr.limbs);
-        online_inputs.extend_from_slice(&lambda_sim_fr.limbs);
-        online_inputs.extend_from_slice(&beta_lambda_sim_fr.limbs);
-        online_inputs.extend_from_slice(&s1_plus_lambda_s2_sim_fr.limbs);
-
-        // Check the constraints
-        assert!(cs.verify_witness(&witness, &online_inputs).is_ok());
-        online_inputs[0].add_assign(&BLSScalar::one());
-        assert!(cs.verify_witness(&witness, &online_inputs).is_err());
     }
 
     #[test]
